Changes to scripting engine. Include stack trace in script error output.
git-svn-id: http://hopmod.googlecode.com/svn/trunk@1728 d7818858-e1b4-11dd-a738-5fe2e3cfbcc5

+++ src/fungu/include/fungu/net/http/connection.hpp
@@ -88,7 +88,7 @@ public:
     void async_send_chunk(const void * data, std::size_t datalen, CompletionHandler handler)
     {
         char buffer[32];
-        std::sprintf(buffer, "%x\r\n", datalen);
+        std::sprintf(buffer, "%x\r\n", static_cast<unsigned int>(datalen));
         
         m_sent_chunk_headers.push(buffer);
